+++ src/config/ConfigKeys.ts
@@ -43,6 +43,7 @@ export const ConfigKeys = {
     sinks: {
       console: 'actor-ts.logger.sinks.console',
       file: 'actor-ts.logger.sinks.file',
+      gelf: 'actor-ts.logger.sinks.gelf',
       otlp: 'actor-ts.logger.sinks.otlp',
     },
   },