+++ src/logging/LoggerFromConfig.ts
@@ -5,6 +5,8 @@ import { ConsoleSink } from './ConsoleSink.js';
 import { isConsoleSinkEnabled, readConsoleSinkOptionsFromConfig } from './ConsoleSinkOptions.js';
 import { FileSink } from './FileSink.js';
 import { isFileSinkEnabled, readFileSinkOptionsFromConfig } from './FileSinkOptions.js';
+import { GelfSink } from './GelfSink.js';
+import { isGelfSinkEnabled, readGelfSinkOptionsFromConfig } from './GelfSinkOptions.js';
 import { OtlpHttpSink } from './OtlpHttpSink.js';
 import { isOtlpSinkEnabled, readOtlpSinkOptionsFromConfig } from './OtlpHttpSinkOptions.js';
 import { parseLogLevel } from './LogLevelName.js';
@@ -55,6 +57,9 @@ export function buildLoggerFromConfig(
   }
   if (isFileSinkEnabled(config)) {
     sinks.push(new FileSink(readFileSinkOptionsFromConfig(config)));
+  }
+  if (isGelfSinkEnabled(config)) {
+    sinks.push(new GelfSink(readGelfSinkOptionsFromConfig(config)));
   }
   if (isOtlpSinkEnabled(config)) {
     sinks.push(new OtlpHttpSink(readOtlpSinkOptionsFromConfig(config)));